+++ src/main/java/com/feiyu/discord/sg/tavern/services/JdaService.java
@@ -31,7 +31,8 @@ public class JdaService {
     private final EventTitleChangeListener eventTitleChangeListener;
     private final EventDetailsMessageListener eventDetailsMessageListener;
     private final NewEventPostListener newEventPostListener;
-    
+    private final EventConfirmationListener eventConfirmationListener;
+
     @Bean
     public net.dv8tion.jda.api.JDA jda(@Value("${discord.bot.token}") String token) throws InterruptedException {
         // Create and configure the JDA instance
@@ -54,6 +55,7 @@ public class JdaService {
                 .addEventListeners(eventTitleChangeListener)
                 .addEventListeners(eventDetailsMessageListener)
                 .addEventListeners(newEventPostListener)
+                .addEventListeners(eventConfirmationListener)
                 .build();
         
         // Optionally wait until the JDA instance is fully ready