+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.5.1] - 2026-05-21
+- beta change for event post to be triggered basis from a button instead of scheduler
+- pilot list selection for new event management flow
+
+## [2.5.0] - 2026-05-19
+- `/pollanonymous` deployed
+
 ## [2.4.1] - 2026-05-14
 - Huge behind the scene changes for invite link controls and tracking